my-feed/requirements.txt
aiogram==3.5.0
fastapi==0.111.0
uvicorn==0.30.0

SQLAlchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

redis==5.0.4
celery==5.4.0

telethon==1.36.0

python-dotenv==1.0.1
pydantic-settings==2.2.1

openai==1.37.0
httpx==0.27.0
